use sst constructs, optimize code for bundling
