+++ frontend/next.config.js
@@ -16,7 +16,12 @@ const nextConfig = {
   },
   rewrites: () => [
     {
-      source: "/api/:path*",
+      source: "/api/lambda/:path*",
+      destination:
+        "https://em9iri9g4j.execute-api.us-west-2.amazonaws.com/:path*",
+    },
+    {
+      source: "/api/training/:path*",
       destination: "http://127.0.0.1:8000/api/:path*",
     },
   ],
