+++ serverless/stacks/AppStack.ts
@@ -1,6 +1,16 @@
-import { Api, StackContext } from "sst/constructs";
+import { Api, Bucket, StackContext } from "sst/constructs";
+import { Bucket as s3Bucket } from "aws-cdk-lib/aws-s3";
 
 export function AppStack({ stack }: StackContext) {
+  const bucket = new Bucket(stack, "dlp-upload-bucket", {
+    cdk: {
+      bucket: s3Bucket.fromBucketArn(
+        stack,
+        "i-dlp-upload-bucket",
+        "arn:aws:s3:::dlp-upload-bucket"
+      ),
+    },
+  });
   const api = new Api(stack, "Api", {
     authorizers: {
       FirebaseAuthorizer: {
@@ -13,24 +23,32 @@ export function AppStack({ stack }: StackContext) {
       },
     },
     defaults: {
+      function: {
+        bind: [bucket],
+      },
       authorizer: "FirebaseAuthorizer",
     },
     routes: {
       "GET /datasets/user/{type}/{filename}/presigned_upload_url":
-        "packages/functions/src/datasets/user_datasets.get_dataset_presigned_upload_url_handler",
+        "packages/functions/src/datasets/user/upload_url.handler",
       "GET /datasets/user/{type}":
-        "packages/functions/src/datasets/user_datasets.get_datasets_handler",
+        "packages/functions/src/datasets/user/all_of_type.handler",
       "GET /datasets/user/{type}/{filename}/columns":
-        "packages/functions/src/datasets/user_datasets.get_dataset_columns_handler",
+        "packages/functions/src/datasets/user/columns.handler",
     },
   });
+
   // Show the API endpoint in the output
   stack.addOutputs({
     ApiEndpoint: api.url,
     GetUserDatasetPresignedUploadUrlFunctionName:
-      api.getFunction("POST /user_datasets/{type}/{filename}")?.functionName ??
-      "",
+      api.getFunction(
+        "GET /datasets/user/{type}/{filename}/presigned_upload_url"
+      )?.functionName ?? "",
     GetUserDatasetsFunctionName:
-      api.getFunction("GET /user_datasets/{type}")?.functionName ?? "",
+      api.getFunction("GET /datasets/user/{type}")?.functionName ?? "",
+    GetUserDatasetColumnsFunctionName:
+      api.getFunction("GET /datasets/user/{type}/{filename}/columns")
+        ?.functionName ?? "",
   });
 }
